hw3/cloudfunction/requirements.txt
google-cloud-storage
google-cloud-logging
google-cloud-pubsub
requests
